Fix coordevents

+++ include/global.fieldmap.h
@@ -77,7 +77,7 @@ struct ObjectEventTemplate
 
 struct WarpEvent
 {
-    s16 x, y;
+    u16 x, y;
     u8 elevation;
     u8 warpId;
     u8 mapNum;
@@ -86,10 +86,10 @@ struct WarpEvent
 
 struct CoordEvent
 {
-    s16 x, y;
+    u16 x, y;
     u8 elevation;
     u16 trigger;
-    u16 index;
+    u8 index;
     u8 *script;
 };
